Remove "testsuites" folder, move contents
