--- project/algorithms/testsuites/__init__.py
@@ -1,3 +1,0 @@
-from .lightning_module_tests import LightningModuleTests
-
-__all__ = ["LightningModuleTests"]
